Combine command classes into cmd.py, pep8 tweaks, README.
