+++ README.md
@@ -1,4 +1,16 @@
 # ff-tool
-Python CLI tool for downloading desktop Firefox version, managing profiles and test prefs
+ff-tool is a Python CLI tool for downloading desktop Firefox versions, managing profiles and test prefs.
+It is largely a glorified convenience wrapper we'ver written around these amazing tools/libraries:
 
-# Work in progress...  DO NOT USE
+* [mozdownload](https://github.com/mozilla/mozdownload)
+* [mozprofile](https://github.com/mozilla/mozprofile)
+
+
+## Example(s)
+```
+$ ff -h
+$ ff download -h
+$ ff download -c nightly
+```
+
+# This tool is work in progress...  DO NOT USE
