+++ configs/channels.ini
@@ -17,4 +17,3 @@ branch=mozilla-aurora
 version=
 type=daily
 branch=mozilla-central
-
