+++ fftool/main.py
@@ -2,10 +2,7 @@
 
 import argparse
 
-from cmd_download import cmd_download
-from cmd_install import cmd_install
-from cmd_profile import cmd_profile
-from cmd_uninstall import cmd_uninstall
+from cmd import CMDDownload, CMDInstall, CMDUninstall, CMDProfile
 
 CHANNELS = ['gr', 'release', 'stable',
             'beta',
@@ -30,21 +27,26 @@ def get_channel(channel):
     return channel
 
 
-parser = argparse.ArgumentParser(prog='ff')
-subparsers = parser.add_subparsers(help='commands', dest='command')
+def main():
+    parser = argparse.ArgumentParser(prog='ff')
+    subparsers = parser.add_subparsers(help='commands', dest='command')
 
-"""
-Global arguments...
-"""
-# parser.add_argument('--foo', action='store_true', help='foo help')
+    """
+    Global arguments...
+    """
+    # parser.add_argument('--foo', action='store_true', help='foo help')
 
-download = cmd_download(subparsers, CHANNELS, DEFAULT_CHANNEL)
-install = cmd_install(subparsers, CHANNELS, DEFAULT_CHANNEL)
-profile = cmd_profile(subparsers)
-uninstall = cmd_uninstall(subparsers, CHANNELS, DEFAULT_CHANNEL)
+    download = CMDDownload(subparsers, CHANNELS, DEFAULT_CHANNEL)
+    install = CMDInstall(subparsers, CHANNELS, DEFAULT_CHANNEL)
+    profile = CMDProfile(subparsers)
+    uninstall = CMDUninstall(subparsers, CHANNELS, DEFAULT_CHANNEL)
 
-options = parser.parse_args()
-if "channel" in options:
-    options.channel = get_channel(options.channel)
+    options = parser.parse_args()
+    if "channel" in options:
+        options.channel = get_channel(options.channel)
 
-options.func(options)
+    options.func(options)
+
+
+if __name__ == '__main__':
+    main()
